+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+* Add #latency to DeliverReport to get how long message took to be delivered.
+
 ## 0.5.0 / 2020-01-21
 
 * Initial release!